Revert premature podspec version bump
Cocoapods spec repo maintainers requested that the version number 
not be incremented.  They prefer to leave the version number at 0.0.1
for all projects that don't have their own semantic version number
scheme using tags in the repo. 

Leaving this at 0.0.1 is fine since the latest change was limited to
the podspec (removing an unnecessary platform restriction).